Add methods to make getting 705 characters easier.

+++ BinUtils/Character.cs
@@ -1,8 +1,8 @@
 ﻿namespace BinUtils;
 
-public record Character(char Char, byte CheckBit, int Value)
+public record Character(char Char, byte CheckBit, byte Value)
 {
-    public int Value { get; set; } = Value;
+    public byte Value { get; set; } = Value;
     public char Char { get; set; } = Char;
     public byte CheckBit { get; init; } = CheckBit;
 }
